org-latex-preview: Option to fine-tune preview opening
* lisp/org-latex-preview.el (org-latex-preview-auto-blacklist,
org-latex-preview-auto--open-this-overlay): Add a blacklist for commands
that should not cause LaTeX preview overlays to be "opened", i.e. for
preview images to be hidden and text revealed when using
`org-latex-preview-auto-mode'.  A common use case would be cursor
navigation commands like `next-line' or `scroll-up-command' that are
often intended to move the view of the buffer, where opening preview
overlays can be distracting.