org-latex-preview: WIP live preview support 3
* lisp/org-latex-preview.el (org-latex-preview-live--update-overlay,
org-latex-preview-live--ensure-overlay,
org-latex-preview-live--src-buffer-setup):
Treat `\[' and `\]' delimited elements as a latex environment for the
purposes of live previews.  This ensures that the live previews for
these are displayed below the fragment instead of inline.
Only update live-previews (directly or via the
`org-latex-preview-update-overlay-functions' hook) when the
`view-text' property of the overlay is true, i.e. when the preview is
"open".  This primarily fixes bugs when editing in the org-src buffer
and using the original org buffer for live previews.
Only setup the org-src buffer for live previews when the source
Org buffer's value of `org-latex-preview-auto-generate' is `live'.