org-latex-preview: move live-preview cache to main cache
* lisp/org-latex-preview.el (org-latex-preview-live--clearout,
org-latex-preview-live--record-hook,
org-latex-preview-live--cache-count,
org-latex-preview-live--last-hash): When exiting an overlay being
live-previewed, hand off data for the last generated image from
the live-preview cache to the main cache, and regenerate the
preview to update the image.  To do this, we track the hash of the
live-previewed fragment in a buffer-local variable
`org-latex-preview-live--last-hash`, taking care to avoid false
positives from renumbering runs.