org-compat: Handle auto in org-create-formula-image
* lisp/org-compat.el (org-create-formula-image): Fix error caused
by setting foreground or background to `auto' in
`org-format-latex-options' (or
`org-latex-preview-appearance-options') and calling
`org-create-formula-image'.

Note that this is not the intended usage of
`org-create-formula-image', the correct way to generate latex
previews with the old Org API is to call `org-format-latex`, which
handles the `auto' color specification.  However some third party
packages (like pdf-tools) do this nevertheless, and are broken now
that `auto' is part of the default value for
`org-latex-preview-appearance-options'.  This fix works around the
issue.