org-latex-preview: Fix wrong fragment caching with live preview
* lisp/org-latex-preview.el (org-latex-preview-live--teardown,
org-latex-preview-live--clearout,
org-latex-preview-live--record-hook): Ensure that the
`org-latex-preview-live--record-hook' only runs when
live-previewing fragments.  This avoids
`org-latex-preview-live--last-hash' being set to the wrong hash
and `org-latex-preview-live--preview-times' recording
non-live-preview times.  When clearing out live previews and
copying from the live-preview cache to the main cache, ensure that
the fragment being copied is actually in hte live-preview cache.