org-latex-preview: Reduce unnecessary work
* lisp/org-latex-preview.el (org-latex-preview--get-preamble): When
calling `org-export--annotate-info', we only need a small part of the
information collected (and work done) is actually relevant.  We reduced
some of the work done by let-binding `org-inhibit-startup' to t and
`org-export-use-babel' to nil, but we were still doing more than needed:
we don't need any form of macro-expansion, link-expansion, or citation
work.  Futhermore, while it is theoretically possible that there might
be a relevant hook added to `org-mode-hook', in practice this seems
extremely unlikely.  As such, it seems sensible to inhibit hooks and see
if any issues crop up.